Rename `apiManager.ts` to `dataUpload.ts`
